feature(botonic-react): use agent image when the message is from an agent and the image exists
